Update Data_Cleaning_Ice_Cream_Data.ipynb
The initial data received in a CSV file  included the columns "Temperature" in Celsius and "Revenue". I wanted to converted temperature from metric to imperial units. So I created a Fahrenheit row in Excel (pictured below). Saved that file and proceeded to my Jupyter Notebook to being the process of analyzing the data with Python.